+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Changed
 - Fixed movies from Kinox.to not working
+- Add streams to stream result list as they load instead of all at once
 
 ## [0.5.0] - 2021-09-02
 ### Added